requirements.txt
# Core dependencies
requests>=2.28.1
pytest>=7.3.1
pyyaml>=6.0
python-dotenv>=1.0.0
jsonschema>=4.17.0

# UI Testing dependencies
selenium>=4.10.0
webdriver-manager>=3.8.6
pillow>=9.5.0
fake-useragent>=1.1.3
pytest-selenium>=4.0.0

# Reporting dependencies
pytest-html>=3.2.0
allure-pytest>=2.13.2
dash>=2.10.0
plotly>=5.14.0
pandas>=2.0.0

# AI dependencies (all optional)
openai>=1.0.0  # Optional: for OpenAI models
google-generativeai>=0.3.0  # Optional: for Google Gemini models

# Notification dependencies
slack_sdk>=3.19.0

# Forecasting and analysis
prophet>=1.1.4  # For time-series forecasting
scikit-learn>=1.2.2  # For ML-based analysis

# CI/CD integration
jenkins-job-builder>=5.0.0  # Optional: for Jenkins integration
